Hide types.h from user program
  * Add stddef.h for `size_t`
  * Add stdint.h for `uintptr_t`

+++ commonsrc/commonsrc.h
@@ -1,13 +1,15 @@
 #pragma once
 
+#include "stddef.h"
+
 // string.c
-int memcmp(const void*, const void*, uint);
-void* memmove(void*, const void*, uint);
-void* memset(void*, int, uint);
+int memcmp(const void*, const void*, size_t);
+void* memmove(void*, const void*, size_t);
+void* memset(void*, int, size_t);
 char* safestrcpy(char*, const char*, int);
 int strlen(const char*);
-int strncmp(const char*, const char*, uint);
-char* strncpy(char*, const char*, int);
+int strncmp(const char*, const char*, size_t);
+char* strncpy(char*, const char*, size_t);
 int strlen(const char*);
 char* strchr(const char*, char);
