+++ include/stdlib.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "types.h"
+#include "stddef.h"
 
 #define EXIT_SUCCESS  (0)
 #define EXIT_FAILURE  (1)
@@ -8,8 +8,8 @@
 int exit(int);
 int atexit(void (*function)(void));
 
-void* malloc(uint);
-void* calloc(uint);
+void* malloc(size_t);
+void* calloc(size_t);
 void free(void*);
 
 int atoi(const char*);
